refacotred and cleaned test using parameters
