+++ tests/test_report_creation/fixtures.py
@@ -1,0 +1,86 @@
+import textwrap
+
+import pytest
+
+
+@pytest.fixture()
+def sample_test(testdir: pytest.Testdir) -> pytest.Testdir:
+    _ = testdir.makefile(
+        ".feature",
+        scenario=textwrap.dedent(
+            """\
+            Feature: Testing
+                Scenario: Test scenario
+                    Given I have a scenario
+                    When I start the test
+                    And I know it will fails
+                    Then It fails
+            """
+        ),
+    )
+
+    _ = testdir.makepyfile(
+        textwrap.dedent(
+            """\
+        import pytest
+        from pytest_bdd import given, when, then, scenario
+
+        @scenario("scenario.feature", "Test scenario")
+        def test_scenario():
+            pass
+
+        @given("I have a scenario")
+        def _():
+            pass
+
+        @when("I start the test")
+        def _():
+            pass
+
+        @when("I know it will fails")
+        def _():
+            pass
+
+        @then('It fails')
+        def _():
+            assert False
+
+        """
+        )
+    )
+
+    return testdir
+
+
+@pytest.fixture()
+def sample_test_with_utf_8_characters(testdir: pytest.Testdir) -> pytest.Testdir:
+    _ = testdir.makefile(
+        ".feature",
+        scenario=textwrap.dedent(
+            """\
+            Feature: Testing utf-8 characters like 👋 and 貓 feature
+                Scenario: Testing utf-8 characters like 👋 and 貓 scenario
+                    Given I include a 👋 character
+                    Given I include a 貓 character
+            """
+        ),
+    )
+
+    _ = testdir.makepyfile(
+        textwrap.dedent(
+            """\
+        import pytest
+        from pytest_bdd import given, when, then, scenario, parsers
+
+        @scenario("scenario.feature", "Testing utf-8 characters like 👋 and 貓 scenario")
+        def test_scenario():
+            pass
+
+        @given(parsers.parse("I include a {character} character"))
+        def _(character):
+            pass
+
+        """
+        )
+    )
+    return testdir
